+++ spec/fixtures/swift/ignored/Owner.swift
@@ -13,8 +13,8 @@ final class Owner: Object {
     case shop = "shop"
   }
 
-  dynamic var name: String?
-  dynamic var shop: Shop?
+  @objc dynamic var name: String?
+  @objc dynamic var shop: Shop?
 
   // Specify properties to ignore (Realm won't persist these)
   override static func ignoredProperties() -> [String] {