+++ spec/fixtures/swift/inverse/Dog.swift
@@ -10,8 +10,8 @@ final class Dog: Object {
     case name = "name"
   }
 
-  dynamic var age: Int16 = 0
-  dynamic var name: String = ""
+  @objc dynamic var age: Int16 = 0
+  @objc dynamic var name: String = ""
 	
   let owners = LinkingObjects(fromType: Person.self, property: "dogs")
   var owner: Person? { return owners.first }