+++ spec/fixtures/swift/realm/FidelityCard.swift
@@ -14,8 +14,8 @@ final class FidelityCard: Object {
     case user = "user"
   }
 
-  dynamic var identifier: Int16 = 0
-  dynamic var points: Int32 = 0
-  dynamic var user: User?
+  @objc dynamic var identifier: Int16 = 0
+  @objc dynamic var points: Int32 = 0
+  @objc dynamic var user: User?
 
 }