+++ spec/fixtures/kotlin/inverse/Dog.kt
@@ -5,7 +5,7 @@ package com.gyro.tests
 import io.realm.RealmObject
 import io.realm.annotations.Required
 
-class Dog : RealmObject() {
+open class Dog : RealmObject() {
 
     object Attributes {
         const val AGE: String = "age"