+++ spec/fixtures/kotlin/primary/FidelityCard.kt
@@ -6,7 +6,7 @@ import io.realm.RealmObject
 import io.realm.annotations.PrimaryKey
 import io.realm.annotations.Required
 
-class FidelityCard : RealmObject() {
+open class FidelityCard : RealmObject() {
 
     object Attributes {
         const val IDENTIFIER: String = "identifier"