+++ spec/fixtures/kotlin/realm/Product.kt
@@ -5,7 +5,7 @@ package com.gyro.tests
 import io.realm.RealmObject
 import io.realm.annotations.Required
 
-class Product : RealmObject() {
+open class Product : RealmObject() {
 
     object Attributes {
         const val BRAND: String = "brand"