+++ spec/fixtures/kotlin/realm/Shop.kt
@@ -6,7 +6,7 @@ import io.realm.RealmList
 import io.realm.RealmObject
 import io.realm.annotations.Required
 
-class Shop : RealmObject() {
+open class Shop : RealmObject() {
 
     object Attributes {
         const val NAME: String = "name"