+++ spec/fixtures/kotlin/realm/User.kt
@@ -7,7 +7,7 @@ import java.util.Date
 import io.realm.RealmObject
 import io.realm.annotations.Required
 
-class User : RealmObject() {
+open class User : RealmObject() {
 
     object Attributes {
         const val BIRTHDAY: String = "birthday"